# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Returns Services API",
        "description": "API documentation for Returns Service App",
        "contact": {
            "email": "admin@adm.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://localhost/api/",
            "description": "Returns App API"
        }
    ],
    "paths": {
...
        "/polaMakans": {
            "get": {
                "tags": [
                    "PolaMakan"
                ],
                "summary": "getPolaMakanList",
                "description": "Get all PolaMakans",
                "operationId": "8cd37763e58b5da6bff16131d5b827dc",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "PolaMakan"
                ],
                "summary": "createPolaMakan",
                "description": "Create PolaMakan",
                "operationId": "a6c026868bd24038d3197547ae35ab99",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "name": {
                                        "description": "desc",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/polaMakans/{id}": {
            "get": {
                "tags": [
                    "PolaMakan"
                ],
                "summary": "getPolaMakanItem",
                "description": "Get PolaMakan",
                "operationId": "ab96e237f682ee1de6f4eb42c7568cab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of PolaMakan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "PolaMakan"
                ],
                "summary": "updatePolaMakan",
                "description": "Update PolaMakan",
                "operationId": "f7f85fcef00533ccb54292926d6b8e66",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of PolaMakan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "name": {
                                        "description": "desc",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "PolaMakan"
                ],
                "summary": "deletePolaMakan",
                "description": "Delete PolaMakan",
                "operationId": "199f774a838beb459c95630c1e590a13",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of PolaMakan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    },
...
}